[kernel] Add library to restrict kernel memory
This library takes memory ranges of base / size along with a
configuration specifying where the kernel and ramdisk were loaded
in memory. Using that information, it checks for a predefined memory
limit being imposed on the system and returns io vectors necessary
to fulfill that requirement, but still maintain space for the kernel
and ramdisk. If the ramdisk is too large to fit within the memory
specified (A problem only seen with Fuchsia builds) then the ramdisk
will be truncated. The library will use the fewest number of vectors
possible so that pmm arenas can be kept to a minimum.

Change-Id: I184db47f9323c5aa22471cd396731a2f83a4729c